+++ src/components/organisms/[module]-daily-lesson/answer-input/index.tsx
@@ -13,6 +13,7 @@ export type BaseAnswerInputProps = AnswerInputProps & {
 };
 
 type AnswerInputProps = {
+  className?: string;
   result?: DLAnswer;
   renderCheckButton: (getCorrectAnswers: () => DLAnswer, disabled: boolean) => JSX.Element;
 };
@@ -44,7 +45,7 @@ const AnswerInput = (props: AnswerInputProps) => {
     case EnumDLContentType.PRONOUNCIATION:
       return (
         <MicrophonePermissionProvider>
-          <Pronounciation />
+          <Pronounciation key={index} />
         </MicrophonePermissionProvider>
       );
     case EnumDLContentType.FILL_IN_THE_BLANK:
