+++ src/components/organisms/question-action-buttons.tsx
@@ -5,9 +5,9 @@ import { useTranslation } from "react-i18next";
 
 import { Button } from "@/components/ui";
 import { useLessonCompletion } from "@/hooks/react-query/use-daily-lesson";
+import useAnswerDailyLesson from "@/hooks/use-answer-daily-lesson";
 import useDailyLessonStore, { DLAnswer } from "@/hooks/zustand/use-daily-lesson-store";
-import { cn, getDuration } from "@/lib/utils";
-import { Route } from "@/routes/_authenticated/daily-lesson/$dailyLessonId";
+import { cn } from "@/lib/utils";
 
 type QuestionActionButtonsProps = {
   getCorrectAnswers: () => DLAnswer;
@@ -18,13 +18,11 @@ const QuestionActionButtons = ({ getCorrectAnswers, disabled }: QuestionActionBu
   const {
     answerQuestion,
     nextQuestion,
-    lessonState: { currentQuestion, isCompleted, startTime },
+    lessonState: { currentQuestion, isCompleted },
     learnerAnswers,
-    setResult,
     setShowExplanation,
   } = useDailyLessonStore();
   const lessonCompletionMutation = useLessonCompletion();
-  const { dailyLessonId } = Route.useParams();
 
   const { t } = useTranslation("question");
   const isAnswered = currentQuestion ? learnerAnswers[currentQuestion.index] !== undefined : false;
@@ -55,42 +53,7 @@ const QuestionActionButtons = ({ getCorrectAnswers, disabled }: QuestionActionBu
     };
   }, [isAnswered, currentQuestion, disabled, lessonCompletionMutation.isPending, isCompleted]);
 
-  useEffect(() => {
-    if (isCompleted) {
-      const statistic = learnerAnswers.reduce(
-        (acc, answer) => {
-          if (answer.correct_letters) {
-            const accCorrectLetters = answer.correct_letters
-              .split("")
-              .filter((letter) => letter === "1").length;
-            return {
-              numberOfCorrect: acc.numberOfCorrect + accCorrectLetters,
-              totalOfQuestions: acc.totalOfQuestions + 1,
-            };
-          }
-          const accCorrect = acc.numberOfCorrect + answer.numberOfCorrect;
-          const accTotal = acc.totalOfQuestions + answer.totalOfQuestions;
-
-          return { numberOfCorrect: accCorrect, totalOfQuestions: accTotal };
-        },
-        { numberOfCorrect: 0, totalOfQuestions: 0 }
-      );
-
-      lessonCompletionMutation.mutate(
-        {
-          lessonId: Number(dailyLessonId),
-          correctAnswers: statistic.numberOfCorrect,
-          wrongAnswers: statistic.totalOfQuestions - statistic.numberOfCorrect,
-          duration: getDuration(startTime),
-        },
-        {
-          onSuccess: (returnData) => {
-            setResult(returnData);
-          },
-        }
-      );
-    }
-  }, [isCompleted]);
+  useAnswerDailyLesson();
 
   useEffect(() => {
     if (isAnswered) {
